Refactored More/Comments Removed
Removed and tested the removal of all print statements throughout the
app, as well as refactored login and logout more. There are still 4
warnings in the console, though 3 are related to the same (inits with
tab bar selected icons), and the other is opening a stream with
NSConnections Causing a warning.